python_agent/requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.32.0
websockets==13.1
requests==2.32.3
pydantic==2.9.2
python-multipart==0.0.12
openai>=1.35.0
websocket-client==1.7.0
python-dotenv==1.0.0
pefile>=2023.2.7
reportlab>=4.0.0
lief>=0.14.0
prettytable>=3.0.0
